% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbxml_handle.R
\name{df2SQLite}
\alias{df2SQLite}
\title{Store drugbank dataframe into SQLite database under \code{extdata} directory of \code{drugbankR} package}
\usage{
df2SQLite(dbdf, version)
}
\arguments{
\item{dbdf}{Drugbank dataframe generated by \code{\link{dbxml2df}} function.}

\item{version}{Character represents drugbank version of the input drugbank dataframe generated by \code{\link{dbxml2df}} function}
}
\value{
Create SQLite database (drugbank_version.db) under \code{extdata} directory of \code{drugbankR} package.
}
\description{
Store specific version of drugbank dataframe into SQLite database under \code{extdata} directory of \code{drugbankR} package
}
\examples{
\dontrun{
## download the original drugbank database (http://www.drugbank.ca/releases/latest) (xml file) 

## into your current directory, and rename as drugbank.xml

drugbank_dataframe = dbxml2df(xmlfile="drugbank.xml", version="5.0.10")

## store the converted drugbank dataframe into SQLite database, 

## The path of the generated SQLite database (drugbank_version.db) is 

## \\code{system.file("extdata", package="drugbankR")}

df2SQLite(dbdf=drugbank_dataframe, version="5.0.10") # set version as version of xml file
}
}
\seealso{
\code{\link{dbxml2df}}
}
\author{
Yuzhu Duan \url{yduan004@ucr.edu}
}
