% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{queryDB}
\alias{queryDB}
\title{Query DrugBank SQLite database.}
\usage{
queryDB(ids = NULL, type, db_path)
}
\arguments{
\item{ids}{character vector, represents DrugBank IDs.}

\item{type}{one of "getAll","getIDs","whichFDA","getTargets" 
\itemize{
  \item getAll: get the entire drugbank dataframe, argument \code{ids} is ignored

  \item getIDs: get all the drugbank ids, argument \code{ids} is ignored

  \item whichFDA: given drugbank ids, determine whether they are FDA approved

  \item getTargets: given drugbank ids, get their targets ids (DrugBank_id, UniProt_id, symbol...).
}}

\item{db_path}{Character(1), path to the DrugBank SQLite database generated
from \code{\link{dbxml2df}} and \code{\link{df2SQLite}} function}
}
\value{
\code{getAll}: drugbank dataframe

\code{getIDs}: character vector of all the drugbank ids

\code{whichFDA}: logical vector

\code{getTargets}: data.frame containing DrugBank_id, UniProt_id, UniProt_name 
and gene_symbol of the targets of the query drugs
}
\description{
This function can be used to

1. get the entire drugbank dataframe

2. get all the drugbank ids

3. given drugbank ids, determine whether the drugs are FDA approved

4. given drugbank ids, get their targets ids (DrugBank_id, UniProt_id, symbol...).
}
